Rename adscc to ADSCitationCapture to comply with blueprint
