+++ .gitignore
@@ -103,3 +103,4 @@ ENV/
 local_config.py
 logs/*
 python/*
+alembic/versions/
