+++ requirements.txt
@@ -1,7 +1,5 @@
-git+https://github.com/adsabs/ADSPipelineUtils.git@master
-SQLAlchemy==1.1.11
+git+https://github.com/adsabs/ADSPipelineUtils.git@v1.0.5
 psycopg2==2.7.1
 alembic==0.9.3
-celery==4.1.0
 sqlalchemy-postgres-copy==0.5.0
 lxml==4.0.0
